user role added controllers

+++ app/Http/Controllers/Api/CronController.php
@@ -7,7 +7,7 @@ use App\Http\Controllers\Controller;
 use App\Models\Cron;
 use Illuminate\Http\Request;
 use Validator;
-
+use App\Permissions\Permission;
 class CronController extends Controller
 {
     /**
@@ -17,8 +17,17 @@ class CronController extends Controller
      */
     public function index()
     {
-        $crons = Cron::all();
-        return response()->json($crons);
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_CRON_INDEX)) {
+            $crons = Cron::all();
+            return response()->json($crons);
+        }
+        else{
+            return response()->json([
+                'error'=> 'Error',
+                'message' => 'Not Authorized.'
+            ]);
+        }
     }
 
     /**
@@ -29,46 +38,57 @@ class CronController extends Controller
      */
     public function store(Request $request)
     {
-        $input = $request->all();
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_CRON_STORE)) {
+            $input = $request->all();
 
-        $validator = Validator::make($input, [
-            'name' => 'required|string|max:50',
-            'action' => 'required|string|max:50',
-            'expression' => 'required|string|max:50',
-            'is_running' => 'required|numeric',
-            'start_time' => 'required|date',
-            'end_time' => 'required|date',
-            'timezone' => 'required|string|max:50',
-            'location' => 'required|string|max:50',
-            'user_id' => 'nullable|uuid'
-        ]);
+            $validator = Validator::make($input, [
+                'name' => 'required|string|max:50',
+                'action' => 'required|string|max:50',
+                'expression' => 'required|string|max:50',
+                'is_running' => 'required|numeric',
+                'start_time' => 'required|date',
+                'end_time' => 'required|date',
+                'timezone' => 'required|string|max:50',
+                'location' => 'required|string|max:50',
+                'user_id' => 'nullable|uuid'
+            ]);
+            
+            if($validator->fails()){
+                return response()->json([
+                    "error" => "Validation Error",
+                    "code"=> 0,
+                    "message"=> $validator->errors()
+                ]);
+            }
+    
+            $input['uniqueid'] = 'sdfasdfasdf'; // must be random string
+    
+            try {
+                $cron = Cron::create($input);
+                return response()->json($cron);
+            } catch (\Exception $e) {
+                if (App::environment('local')) {
+                    $message = $e->getMessage();
+                }
+                else{
+                    $message = "Cron store error";
+                }
+                return response()->json([
+                    "error" => "Error",
+                    "code"=> 0,
+                    "message"=> $message
+                ]);
+            }
+        }
+        else{
+            return response()->json([
+                'error'=> 'Error',
+                'message' => 'Not Authorized.'
+            ]);
+        }
+
         
-        if($validator->fails()){
-            return response()->json([
-                "error" => "Validation Error",
-                "code"=> 0,
-                "message"=> $validator->errors()
-            ]);
-        }
-
-        $input['uniqueid'] = 'sdfasdfasdf'; // must be random string
-
-        try {
-            $cron = Cron::create($input);
-            return response()->json($cron);
-        } catch (\Exception $e) {
-            if (App::environment('local')) {
-                $message = $e->getMessage();
-            }
-            else{
-                $message = "Cron store error";
-            }
-            return response()->json([
-                "error" => "Error",
-                "code"=> 0,
-                "message"=> $message
-            ]);
-        }
     }
 
     /**
@@ -79,8 +99,16 @@ class CronController extends Controller
      */
     public function show(Cron $cron)
     {
-        //
-        return response()->json($cron);
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_CRON_SHOW)) {
+            return response()->json($cron);
+        }
+        else{
+            return response()->json([
+                'error'=> 'Error',
+                'message' => 'Not Authorized.'
+            ]);
+        }
     }
 
     /**
@@ -92,44 +120,53 @@ class CronController extends Controller
      */
     public function update(Request $request, Cron $cron)
     {
-        $input = $request->all();
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_CRON_UPDATE)) {
+            $input = $request->all();
 
-        $validator = Validator::make($input, [
-            'name' => 'nullable|string|max:50',
-            'action' => 'nullable|string|max:50',
-            'expression' => 'nullable|string|max:50',
-            'is_running' => 'nullable|numeric',
-            'start_time' => 'nullable|date',
-            'end_time' => 'nullable|date',
-            'timezone' => 'nullable|string|max:50',
-            'location' => 'nullable|string|max:50',
-            'user_id' => 'nullable|uuid'
-        ]);
-
-        if($validator->fails()){
-            return response()->json([
-                "error" => "Validation Error",
-                "code"=> 0,
-                "message"=> $validator->errors()
+            $validator = Validator::make($input, [
+                'name' => 'nullable|string|max:50',
+                'action' => 'nullable|string|max:50',
+                'expression' => 'nullable|string|max:50',
+                'is_running' => 'nullable|numeric',
+                'start_time' => 'nullable|date',
+                'end_time' => 'nullable|date',
+                'timezone' => 'nullable|string|max:50',
+                'location' => 'nullable|string|max:50',
+                'user_id' => 'nullable|uuid'
             ]);
+    
+            if($validator->fails()){
+                return response()->json([
+                    "error" => "Validation Error",
+                    "code"=> 0,
+                    "message"=> $validator->errors()
+                ]);
+            }
+    
+            try {
+                $cron->update($input);
+                return response()->json($cron);
+            } catch (\Exception $e) {
+                if (App::environment('local')) {
+                    $message = $e->getMessage();
+                }
+                else{
+                    $message = "Cron update error";
+                }
+                return response()->json([
+                    "error" => "Error",
+                    "code"=> 0,
+                    "message"=> $message
+                ]);
+            }
         }
-
-        try {
-            $cron->update($input);
-            return response()->json($cron);
-        } catch (\Exception $e) {
-            if (App::environment('local')) {
-                $message = $e->getMessage();
-            }
-            else{
-                $message = "Cron update error";
-            }
+        else{
             return response()->json([
-                "error" => "Error",
-                "code"=> 0,
-                "message"=> $message
+                'error'=> 'Error',
+                'message' => 'Not Authorized.'
             ]);
-        }
+        }   
     }
 
     /**
@@ -140,7 +177,16 @@ class CronController extends Controller
      */
     public function destroy(Cron $cron)
     {
-        $cron->delete();
-        return response()->json();
+        $user = Auth::user();
+        if ($user->tokenCan(Permission::CAN_ALL) || $user->tokenCan(Permission::CAN_CRON_DESTROY)) {
+            $cron->delete();
+            return response()->json();
+        }
+        else{
+            return response()->json([
+                'error'=> 'Error',
+                'message' => 'Not Authorized.'
+            ]);
+        }
     }
 }
